papers/s8_uha_ssot_oct2025/requirements.txt
# minimal runtime for fetch/parse + basic stats
numpy>=1.26
scipy>=1.11
pandas>=2.0
pyyaml>=6.0
requests>=2.31
tqdm>=4.66
matplotlib>=3.8
